+++ YarnSpinner.Compiler/Compiler.cs
@@ -418,6 +418,12 @@ namespace Yarn.Compiler
         public IEnumerable<Diagnostic> Diagnostics { get; internal set; }
 
         /// <summary>
+        /// Gets the collection of <see cref="DebugInfo"/> objects for each node
+        /// in <see cref="Program"/>.
+        /// </summary>
+        public IReadOnlyDictionary<string, DebugInfo> DebugInfo { get; internal set; }
+
+        /// <summary>
         /// Combines multiple <see cref="CompilationResult"/> objects together
         /// into one object.
         /// </summary>
@@ -433,6 +439,7 @@ namespace Yarn.Compiler
             var declarations = new List<Declaration>();
             var tags = new Dictionary<string, IEnumerable<string>>();
             var diagnostics = new List<Diagnostic>();
+            var nodeDebugInfos = new Dictionary<string, DebugInfo>();
 
             foreach (var result in results)
             {
@@ -455,6 +462,14 @@ namespace Yarn.Compiler
                 {
                     diagnostics.AddRange(result.Diagnostics);
                 }
+
+                if (result.DebugInfo != null)
+                {
+                    foreach (var kvp in result.DebugInfo)
+                    {
+                        nodeDebugInfos.Add(kvp.Key, kvp.Value);
+                    }
+                }
             }
 
             return new CompilationResult
@@ -462,10 +477,95 @@ namespace Yarn.Compiler
                 Program = Program.Combine(programs.ToArray()),
                 StringTable = stringTableManager.StringTable,
                 Declarations = declarations,
+                DebugInfo = nodeDebugInfos,
                 ContainsImplicitStringTags = stringTableManager.ContainsImplicitStringTags,
                 FileTags = tags,
                 Diagnostics = diagnostics,
             };
+        }
+    }
+
+    /// <summary>
+    /// Contains debug information for a node in a Yarn file.
+    /// </summary>
+    public class DebugInfo
+    {
+        /// <summary>
+        /// Gets or sets the file that this DebugInfo was produced from.
+        /// </summary>
+        internal string FileName { get; set; }
+
+        /// <summary>
+        /// Gets or sets the node that this DebugInfo was produced from.
+        /// </summary>
+        internal string NodeName { get; set; }
+
+        /// <summary>
+        /// Gets or sets the mapping of instruction numbers to line and
+        /// character information in the file indicated by <see
+        /// cref="FileName"/>.
+        /// </summary>
+        internal Dictionary<int, (int Line, int Character)> LineInfos { get; set; } = new Dictionary<int, (int Line, int Character)>();
+
+        /// <summary>
+        /// Gets a <see cref="LineInfo"/> object that describes the specified
+        /// instruction at the index <paramref name="instructionNumber"/>.
+        /// </summary>
+        /// <param name="instructionNumber">The index of the instruction to
+        /// retrieve information for.</param>
+        /// <returns>A <see cref="LineInfo"/> object that describes the position
+        /// of the instruction.</returns>
+        /// <exception cref="ArgumentOutOfRangeException">Thrown when <paramref
+        /// name="instructionNumber"/> is less than zero, or greater than the
+        /// number of instructions present in the node.</exception>
+        public LineInfo GetLineInfo(int instructionNumber)
+        {
+            if (this.LineInfos.TryGetValue(instructionNumber, out var info))
+            {
+                return new LineInfo
+                {
+                    FileName = this.FileName,
+                    NodeName = this.NodeName,
+                    LineNumber = info.Line,
+                    CharacterNumber = info.Character,
+                };
+            }
+            else
+            {
+                throw new ArgumentOutOfRangeException(nameof(instructionNumber));
+            }
+        }
+
+        /// <summary>
+        /// Contains positional information about an instruction.
+        /// </summary>
+        public struct LineInfo
+        {
+            /// <summary>
+            /// The file name of the source that this intruction was produced
+            /// from.
+            /// </summary>
+            public string FileName;
+
+            /// <summary>
+            /// The node name of the source that this intruction was produced
+            /// from.
+            /// </summary>
+            public string NodeName;
+
+            /// <summary>
+            /// The zero-indexed line number in <see cref="FileName"/> that
+            /// contains the statement or expression that this line was produced
+            /// from.
+            /// </summary>
+            public int LineNumber;
+
+            /// <summary>
+            /// The zero-indexed character number in <see cref="FileName"/> that
+            /// contains the statement or expression that this line was produced
+            /// from.
+            /// </summary>
+            public int CharacterNumber;
         }
     }
 
@@ -487,6 +587,12 @@ namespace Yarn.Compiler
         internal Node CurrentNode { get; private set; }
 
         /// <summary>
+        /// Gets the current debug information that describes <see
+        /// cref="CurrentNode"/>.
+        /// </summary>
+        private DebugInfo CurrentDebugInfo { get; set; }
+
+        /// <summary>
         /// Gets or sets a value indicating whether we are currently parsing the
         /// current node as a 'raw text' node, or as a fully syntactic node.
         /// </summary>
@@ -497,6 +603,8 @@ namespace Yarn.Compiler
         /// Gets the program being generated by the compiler.
         /// </summary>
         internal Program Program { get; private set; }
+
+        internal IList<DebugInfo> DebugInfos { get; private set; } = new List<DebugInfo>();
 
         internal FileParseResult fileParseResult { get; private set; }
 
@@ -787,9 +895,17 @@ namespace Yarn.Compiler
             compiler.VariableDeclarations = variableDeclarations;
             compiler.Compile();
 
+            var debugInfoDictionary = new Dictionary<string, DebugInfo>();
+
+            foreach (var debugInfo in compiler.DebugInfos)
+            {
+                debugInfoDictionary.Add(debugInfo.NodeName, debugInfo);
+            }
+
             return new CompilationResult
             {
                 Program = compiler.Program,
+                DebugInfo = debugInfoDictionary,
                 StringTable = stringTableManager.StringTable,
                 ContainsImplicitStringTags = stringTableManager.ContainsImplicitStringTags,
                 Diagnostics = compiler.Diagnostics,
@@ -983,32 +1099,46 @@ namespace Yarn.Compiler
         /// cref="Program" />.
         /// </summary>
         /// <param name="node">The node to append instructions to.</param>
+        /// <param name="debugInfo">The <see cref="DebugInfo"/> object to add
+        /// line debugging information to.</param>
+        /// <param name="sourceLine">The zero-indexed line in the source input
+        /// corresponding to this instruction.</param>
+        /// <param name="sourceCharacter">The zero-indexed character in the
+        /// source input corresponding to this instruction.</param>
         /// <param name="code">The opcode of the instruction.</param>
         /// <param name="operands">The operands to associate with the
         /// instruction.</param>
-        void Emit(Node node, OpCode code, params Operand[] operands)
+        void Emit(Node node, DebugInfo debugInfo, int sourceLine, int sourceCharacter, OpCode code, params Operand[] operands)
         {
             var instruction = new Instruction
             {
-                Opcode = code
+                Opcode = code,
             };
 
             instruction.Operands.Add(operands);
+
+            debugInfo.LineInfos.Add(node.Instructions.Count, (sourceLine, sourceCharacter));
 
             node.Instructions.Add(instruction);
         }
 
         /// <summary>
-        /// Creates a new instruction, and appends it to the current node
-        /// in the <see cref="Program"/>. Called by instances of <see
-        /// cref="CodeGenerationVisitor"/> while walking the parse tree.
+        /// Creates a new instruction, and appends it to the current node in the
+        /// <see cref="Program"/>.
         /// </summary>
+        /// <remarks>
+        /// Called by instances of <see
+        /// cref="CodeGenerationVisitor"/> while walking the parse tree.
+        /// </remarks>
         /// <param name="code">The opcode of the instruction.</param>
+        /// <param name="startToken">The first token in the expression or
+        /// statement that was responsible for emitting this
+        /// instruction.</param>
         /// <param name="operands">The operands to associate with the
         /// instruction.</param>
-        internal void Emit(OpCode code, params Operand[] operands)
+        internal void Emit(OpCode code, IToken startToken, params Operand[] operands)
         {
-            Emit(this.CurrentNode, code, operands);
+            Emit(this.CurrentNode, this.CurrentDebugInfo, startToken?.Line - 1 ?? -1, startToken?.Column ?? -1, code, operands);
         }
 
         /// <summary>
@@ -1052,6 +1182,7 @@ namespace Yarn.Compiler
         public override void EnterNode(YarnSpinnerParser.NodeContext context)
         {
             CurrentNode = new Node();
+            CurrentDebugInfo = new DebugInfo();
             RawTextNode = false;
         }
 
@@ -1060,6 +1191,12 @@ namespace Yarn.Compiler
         public override void ExitNode(YarnSpinnerParser.NodeContext context)
         {
             Program.Nodes[CurrentNode.Name] = CurrentNode;
+
+            CurrentDebugInfo.NodeName = CurrentNode.Name;
+            CurrentDebugInfo.FileName = fileParseResult.Name;
+
+            DebugInfos.Add(CurrentDebugInfo);
+
             CurrentNode = null;
             RawTextNode = false;
         }
@@ -1143,7 +1280,7 @@ namespace Yarn.Compiler
         public override void ExitBody(YarnSpinnerParser.BodyContext context)
         {
             // We have exited the body; emit a 'stop' opcode here.
-            Emit(CurrentNode, OpCode.Stop);
+            Emit(CurrentNode, CurrentDebugInfo, context.Stop.Line - 1, 0, OpCode.Stop);
         }
 
         /// <summary>