+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added `DeclarationBuilder` and `FunctionTypeBuilder` classes. These classes allow external libraries to construct new `Declaration` and `FunctionType` objects, without having to have access to the internal setters.
+- `CompilationResult.DebugInfo` now provides per-instruction positional debug information.
+  - This allows users of the `Compiler` class to access positional information for each instruction, which is an important first step for source-level debugging.
 
 ### Changed